상세페이지 및 서비스

+++ sbbb/src/main/java/com/mysite/sbbb/question/dao/QuestionRepository.java
@@ -3,5 +3,6 @@ package com.mysite.sbbb.question.dao;
 import com.mysite.sbbb.question.domain.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
+
 }
